added node and bucket icons
Printing those means that users must turn on print
backgrounds and/or images. We need the backgrounds
setting anyway for status fields of tables.

If we do not want the requirement, I can make the
print stylesheet not display the images and turn
status text a color vs. using a background color.

Change-Id: I3340efadd443c1b15febc9ad9f2667828c4b1c39