Add fine grained analysis for cache miss ratio
Since cache miss ratio tends to be volatile, it is quite misleading
to just use averge value for a time span, say, one day samplings.
Here we use an approach to find many abnormal segments and correlate
them to current items and active residential ratio. If all of them
are above thresholds, the abnormal segment will be marked as a
questionable one. As such, a time notion is introduced.